+++ tests/testthat.R
@@ -1,4 +1,4 @@
-library("tna")
+suppressPackageStartupMessages(library("tna"))
 library("testthat")
 
 test_check("tna")